CHAT-11 Refactored folder structure and imports.
